feat: almost all typing errors fixed, nano issue remains

+++ src/db_manager.ts
@@ -12,11 +12,11 @@ const dbProtocol = DB_CONNECTION_STRING.split(':')[0];
 
 const mongoClient =
   dbProtocol === 'mongodb' ? new MongoClient(DB_CONNECTION_STRING) : null;
-const mongoDb = dbProtocol === 'mongodb' ? client.db() : null;
+const mongoDb = (dbProtocol === 'mongodb' && mongoClient) ? mongoClient.db() : null;
 const nanoDb = dbProtocol === 'mongodb' ? null : Nano(DB_CONNECTION_STRING);
 
 async function getNextSequence(collectionName: string): Promise<number> {
-  if (dbProtocol === 'mongodb') {
+  if (dbProtocol === 'mongodb' && mongoDb) {
     const ret = await mongoDb.command({
       findAndModify: 'counters',
       query: { _id: collectionName },
@@ -25,7 +25,7 @@ async function getNextSequence(collectionName: string): Promise<number> {
       upsert: true
     });
     return ret.value.seq;
-  } else {
+  } else if (nanoDb) {
     const counterDocId = `counter_${collectionName}`;
     let counterDoc;
 
@@ -39,73 +39,75 @@ async function getNextSequence(collectionName: string): Promise<number> {
     await nanoDb.insert(counterDoc);
     return counterDoc.value;
   }
+  return -1;
 }
 
 const dbManager = {
   async connect(): Promise<void> {
-    if (dbProtocol === 'mongodb') await mongoClient.connect();
+    if (dbProtocol === 'mongodb' && mongoClient) await mongoClient.connect();
   },
 
   async disconnect(): Promise<void> {
-    if (dbProtocol === 'mongodb') await mongoClient.close();
+    if (dbProtocol === 'mongodb' && mongoClient) await mongoClient.close();
   },
 
   /** Get all productions from the database in reverse natural order, limited by the limit parameter */
   async getProductions(limit: number, offset: number): Promise<Production[]> {
     let productions: Production[] = [];
-    if (dbProtocol === 'mongodb') {
-      productions = await mongoDb
+    if (dbProtocol === 'mongodb' && mongoDb) {
+      productions = (await mongoDb
         .collection('productions')
         .find()
         .sort({ $natural: -1 })
         .skip(offset)
         .limit(limit)
-        .toArray();
-    } else {
+        .toArray()) as unknown as Production[];
+    } else if (nanoDb) {
       const response = await nanoDb.list({
         limit: limit,
         skip: offset,
         sort: [{ name: 'desc' }],
         include_docs: true
       });
-      response.rows.forEach((row) => {
+      response.rows.forEach((row: any) => {
         if (row.doc._id.toLowerCase().indexOf('counter') === -1)
           productions.push(row.doc);
       });
     }
-
     return productions as any as Production[];
   },
 
   async getProductionsLength(): Promise<number> {
-    if (dbProtocol === 'mongodb') {
+    if (dbProtocol === 'mongodb' && mongoDb) {
       return await mongoDb.collection('productions').countDocuments();
-    } else {
+    } else if (nanoDb) {
       const productions = await nanoDb.list({ include_docs: false });
       return productions.rows.length;
     }
+    return 0;
   },
 
   async getProduction(id: number): Promise<Production | undefined> {
-    if (dbProtocol === 'mongodb') {
+    if (dbProtocol === 'mongodb' && mongoDb) {
       return mongoDb.collection('productions').findOne({ _id: id as any }) as
         | any
         | undefined;
-    } else {
+    } else if (nanoDb) {
       const production = await nanoDb.get(id.toString());
       return production as any | undefined;
     }
+    return undefined;
   },
 
   async updateProduction(
     production: Production
   ): Promise<Production | undefined> {
-    if (dbProtocol === 'mongodb') {
+    if (dbProtocol === 'mongodb' && mongoDb) {
       const result = await mongoDb
         .collection('productions')
         .updateOne({ _id: production._id as any }, { $set: production });
       return result.modifiedCount === 1 ? production : undefined;
-    } else {
+    } else if (nanoDb) {
       const existingProduction = await nanoDb.get(production._id.toString());
       const updatedProduction = { ...existingProduction, ...production };
       const response = await nanoDb.insert(updatedProduction);
@@ -119,27 +121,27 @@ const dbManager = {
       throw new Error('Failed to get next sequence');
     }
     const production = { name, lines, _id };
-    if (dbProtocol === 'mongodb') {
+    if (dbProtocol === 'mongodb' && mongoDb) {
       await mongoDb.collection('productions').insertOne(production as any);
-      return production;
-    } else {
-      const production = { name, lines, _id };
+    } else if (nanoDb) {
       const response = await nanoDb.insert(production);
-      return response.ok ? production : undefined;
+      if (!response.ok) throw new Error('Failed to insert production');
     }
+    return production;
   },
 
   async deleteProduction(productionId: number): Promise<boolean> {
-    if (dbProtocol === 'mongodb') {
+    if (dbProtocol === 'mongodb' && mongoDb) {
       const result = await mongoDb
         .collection('productions')
         .deleteOne({ _id: productionId as any });
       return result.deletedCount === 1;
-    } else {
+    } else if (nanoDb) {
       const production = await nanoDb.get(productionId.toString());
       const response = await nanoDb.destroy(production._id, production._rev);
       return response.ok;
     }
+    return false;
   },
 
   async setLineConferenceId(
@@ -155,14 +157,14 @@ const dbManager = {
       `Line with id "${lineId}" does not exist for production with id "${productionId}"`
     );
     line.smbConferenceId = conferenceId;
-    if (dbProtocol === 'mongodb') {
+    if (dbProtocol === 'mongodb' && mongoDb) {
       await mongoDb
         .collection('productions')
         .updateOne(
           { _id: productionId as any },
           { $set: { lines: production.lines } }
         );
-    } else {
+    } else if (nanoDb) {
       const existingProduction = await nanoDb.get(productionId.toString());
       const updatedProduction = {
         ...existingProduction,
